[Python] Fix TypeInfo.__hash__ to return native int for Python 3.14
TypeInfo.__hash__ called combine_hash_codes() which returns int32
(Fable's Int32, not a subclass of native int) and passed it through.
Python 3.14 strictly requires __hash__ to return a native int. Wrap the
result with int(), matching the pattern in HashableBase.__hash__.

combine_hash_codes and array_hash stay as int32 since they serve
GetHashCode (which returns int32 by design).

+++ src/fable-library-py/fable_library/reflection.py
@@ -6,7 +6,7 @@ from dataclasses import dataclass
 from typing import Any, cast
 
 from .array_ import Array
-from .core import FSharpRef
+from .core import FSharpRef, int32
 from .record import Record
 from .types import IntegerTypes
 from .union import Union
@@ -55,9 +55,9 @@ class TypeInfo:
         return equals(self, other)
 
     def __hash__(self) -> int:
-        hashes: list[int] = [hash(x) for x in self.generics or []]
-        hashes.append(hash(self.fullname))
-        return combine_hash_codes(hashes)
+        hashes: list[int32] = [int32(hash(x)) for x in self.generics or []]
+        hashes.append(int32(hash(self.fullname)))
+        return int(combine_hash_codes(hashes))
 
 
 def class_type(
